+++ includes/class-tmsm-woocommerce-booking-thalasso.php
@@ -106,6 +106,7 @@ class Tmsm_Woocommerce_Booking_Thalasso {
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-bookroombuttonlabel.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-accommodationurl.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-packageratestable.php';
+			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-packageratescards.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-bookpackageprice.php';
 		}